+++ webpack.config.js
@@ -1,6 +1,7 @@
 const path = require('path');
 module.exports = {
     entry: "./src/index.js",
+    devtool: 'inline-source-map',
     output: {
         filename: "bundle.js",
         path: path.join(__dirname,'dist')
